requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.3.0
pyttsx3==2.90
SpeechRecognition==3.10.0
requests==2.31.0
Pillow==10.0.1 